extractor/requirements.txt
pdfplumber>=0.11
pymupdf>=1.24
pyyaml>=6.0
pytest>=8.0
rapidfuzz>=3.0
openpyxl>=3.1
